Move imageWithDigest to bundle package and make it public
