+++ driver/kubernetes/kubernetes.go
@@ -14,9 +14,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/docker/distribution/reference"
 	"github.com/hashicorp/go-multierror"
-	"github.com/opencontainers/go-digest"
 	"github.com/pkg/errors"
 	batchv1 "k8s.io/api/batch/v1"
 	v1 "k8s.io/api/core/v1"
@@ -33,7 +31,6 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 
-	"github.com/cnabio/cnab-go/bundle"
 	"github.com/cnabio/cnab-go/driver"
 )
 
@@ -363,7 +360,7 @@ func (k *Driver) Run(op *driver.Operation) (driver.OperationResult, error) {
 			},
 		},
 	}
-	img, err := imageWithDigest(op.Image)
+	img, err := op.Image.WithDigest()
 	if err != nil {
 		return driver.OperationResult{}, err
 	}
@@ -713,34 +710,4 @@ func homeDir() string {
 		return h
 	}
 	return os.Getenv("USERPROFILE") // windows
-}
-
-func imageWithDigest(img bundle.InvocationImage) (string, error) {
-	// img.Image can be just the name, name:tag or name@digest
-	ref, err := reference.ParseNormalizedNamed(img.Image)
-	if err != nil {
-		return "", errors.Wrapf(err, "could not parse %s as an OCI reference", img.Image)
-	}
-
-	var d digest.Digest
-	if v, ok := ref.(reference.Digested); ok {
-		// Check that the digests match since it's provided twice
-		if img.Digest != "" && img.Digest != v.Digest().String() {
-			return "", errors.Errorf("The digest %s for the image %s doesn't match the one specified in the image", img.Digest, img.Image)
-		}
-		d = v.Digest()
-	} else if img.Digest != "" {
-		d, err = digest.Parse(img.Digest)
-		if err != nil {
-			return "", errors.Wrapf(err, "invalid digest %s specified for invocation image %s", img.Digest, img.Image)
-		}
-	}
-
-	// Digest was not supplied anywhere
-	if d == "" {
-		return img.Image, nil
-	}
-
-	digestedRef, err := reference.WithDigest(ref, d)
-	return reference.FamiliarString(digestedRef), errors.Wrapf(err, "invalid image digest %s", d.String())
 }
